requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
